+++ lib/go-tc/alerts.go
@@ -75,22 +75,6 @@ func GetHandleErrorsFunc(w http.ResponseWriter, r *http.Request) func(status int
 	}
 }
 
-func HandleErrorsWithType(errs []error, errType ApiErrorType, handleErrs func(status int, errs ...error)) {
-	switch errType {
-	case SystemError:
-		handleErrs(http.StatusInternalServerError, errs...)
-	case DataConflictError:
-		handleErrs(http.StatusBadRequest, errs...)
-	case DataMissingError:
-		handleErrs(http.StatusNotFound, errs...)
-	case ForbiddenError:
-		handleErrs(http.StatusForbidden, errs...)
-	default:
-		log.Errorf("received unknown ApiErrorType from read: %s\n", errType.String())
-		handleErrs(http.StatusInternalServerError, errs...)
-	}
-}
-
 func (alerts *Alerts) ToStrings() []string {
 	alertStrs := []string{}
 	for _, alrt := range alerts.Alerts {