+++ traffic_ops/testing/api/v14/withobjs.go
@@ -52,7 +52,9 @@ const (
 	ProfileParameters
 	Regions
 	Roles
+	ServerCapabilities
 	ServerChecks
+	ServerServerCapabilities
 	Servers
 	Statuses
 	StaticDNSEntries
@@ -86,7 +88,9 @@ var withFuncs = map[TCObj]TCObjFuncs{
 	ProfileParameters:              {CreateTestProfileParameters, DeleteTestProfileParameters},
 	Regions:                        {CreateTestRegions, DeleteTestRegions},
 	Roles:                          {CreateTestRoles, DeleteTestRoles},
+	ServerCapabilities:             {CreateTestServerCapabilities, DeleteTestServerCapabilities},
 	ServerChecks:                   {CreateTestServerChecks, DeleteTestServerChecks},
+	ServerServerCapabilities:       {CreateTestServerServerCapabilities, DeleteTestServerServerCapabilities},
 	Servers:                        {CreateTestServers, DeleteTestServers},
 	Statuses:                       {CreateTestStatuses, DeleteTestStatuses},
 	StaticDNSEntries:               {CreateTestStaticDNSEntries, DeleteTestStaticDNSEntries},